Refactor dualsense, many under the hood improvements
