+++ .gitignore
@@ -1,5 +1,6 @@
 .vscode
 dist
+lib
 node_modules
 app.settings.json
 .env
