+++ packages/controllers/dualsense/package.json
@@ -1,0 +1,23 @@
+{
+    "name": "@lbc/dualsense",
+    "version": "0.2.2-alpha",
+    "license": "MIT",
+    "scripts": {
+        "build": "yarn tsc",
+        "start": "node dist/index.js",
+        "dev": "yarn ts-node src/index.ts"
+    },
+    "devDependencies": {
+        "@types/express": "^4.17.17",
+        "@types/node": "^18.13.0",
+        "prettier": "^2.8.4",
+        "ts-node": "^10.9.1",
+        "typescript": "^4.9.5"
+    },
+    "dependencies": {
+        "axios": "^1.3.3",
+        "dotenv": "^16.0.3",
+        "dualsense-ts": "^5.0.120",
+        "node-hid": "^2.1.2"
+    }
+}
