+++ packages/controllers/web/tsconfig.json
@@ -1,0 +1,6 @@
+{
+    "extends": "../../../tsconfig.json",
+    "compilerOptions": {
+        "jsx": "react"
+    }
+}
