+++ packages/core/app/package.json
@@ -1,0 +1,31 @@
+{
+    "name": "@lbc/app",
+    "version": "0.2.2-alpha",
+    "license": "MIT",
+    "scripts": {
+        "build": "yarn tsc",
+        "start": "node dist/index.js",
+        "dev": "yarn ts-node src/index.ts"
+    },
+    "devDependencies": {
+        "@serialport/list": "^10.5.0",
+        "@types/express": "^4.17.17",
+        "@types/node": "^18.13.0",
+        "@types/uuid": "^9.0.0",
+        "prettier": "^2.8.4",
+        "ts-node": "^10.9.1",
+        "typescript": "^4.9.5"
+    },
+    "dependencies": {
+        "@serialport/bindings-interface": "^1.2.2",
+        "axios": "^1.3.3",
+        "cors": "^2.8.5",
+        "dotenv": "^16.0.3",
+        "express": "^4.18.2",
+        "serialport": "^10.5.0",
+        "socket.io": "^4.6.0",
+        "sqlite": "^4.1.2",
+        "sqlite3": "^5.1.4",
+        "uuid": "^9.0.0"
+    }
+}
