+++ packages/core/app/src/serial/serialUtil.ts
